Strategy-Backtester-main/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
alpha_vantage>=2.3.1
python-dotenv>=1.0.0
streamlit>=1.30.0 